🎨	: 패키지명 변경 +  Member Id (Long) -> (String)
