+++ src/main/java/org/kcsmini2/ojeommo/member/entity/Member.java
@@ -1,0 +1,75 @@
+package org.kcsmini2.ojeommo.member.entity;
+
+
+import jakarta.persistence.*;
+import lombok.*;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Member implements UserDetails {
+
+    @Id
+    private String id;
+
+    @Column
+    private String pw;
+
+    @Column
+    private String nickname;
+
+    @Column
+    private String name;
+
+    @Column
+    private String email;
+
+    @Builder
+    public Member(String id, String pw, String nickname, String name, String email) {
+        this.id = id;
+        this.pw = pw;
+        this.nickname = nickname;
+        this.name = name;
+        this.email = email;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return pw;
+    }
+
+    @Override
+    public String getUsername() {
+        return name;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return false;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return false;
+    }
+}
