+++ src/main/java/org/kcsmini2/ojeommo/member/entity/PartyPk.java
@@ -1,0 +1,11 @@
+package org.kcsmini2.ojeommo.member.entity;
+
+import java.io.Serializable;
+
+
+public class PartyPk implements Serializable {
+
+    private String memberId;
+
+    private Long boardId;
+}
