--- src/main/java/org/kcsmini2/ojummo/category/entity/Category.java
@@ -1,22 +1,0 @@
-package org.kcsmini2.ojummo.category.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Column
-    private String category;
-
-    @Builder
-    public Category(String category) {
-        this.category = category;
-    }
-}
