--- src/main/java/org/kcsmini2/ojummo/member/entity/PartyPk.java
@@ -1,11 +1,0 @@
-package org.kcsmini2.ojummo.member.entity;
-
-import java.io.Serializable;
-
-
-public class PartyPk implements Serializable {
-
-    private Long memberId;
-
-    private Long boardId;
-}
