+++ src/test/java/org/kcsmini2/ojeommo/OjummoApplicationTests.java
@@ -1,4 +1,4 @@
-package org.kcsmini2.ojummo;
+package org.kcsmini2.ojeommo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
